requirements.txt
discord.py[voice]>=1.4.1
requests
discord-buttons-plugin
